logging

+++ local_folder_service.py
@@ -21,43 +21,44 @@ class OneThreadProcessor:
             files = [os.path.join(self.input, file_name) for file_name in os.listdir(self.input)]
             # fetch files using creation date
             sorted_files = sorted(files, key=lambda x: os.path.getctime(x))
-            t = time.time_ns()
-            for input_file_name in sorted_files:
-                # process in place
-                _name = os.path.split(input_file_name)[-1]
-                # use same names in output as in input
-                target_name = os.path.join(self.processor.tts.folder, os.path.splitext(_name)[0] + ".wav")
-                if _name == "newuser":
-                    # initiate new protocol, ask user his name
-                    self.username = None
-                    output_file_name = self.processor.tts.get_audio("Здравствуйте, меня зовут Сергей Капица! Представьтесь, пожалуйста.", output_name=target_name)
-                else:
-                    if self.username is None: # expect user name as an answer
-                        user_answer = self.processor.asr.get_text(input_file_name)
-                        if user_answer is None:
-                            print(f"Couldn't fetch username: {user_answer} from file {input_file_name}. Setting name to Дорогой друг")
-                            self.username = "Дорогой друг"
-                        else:
-                            user_answer = user_answer.strip(".,! ").capitalize()
-                            print(f"User answered: {user_answer}")
-                            self.username = user_answer
-                        # self.processor.llm.set_engine(user_name=None, reset=True, custom_system_prompt="""Ты - система аутентификации для ASR. Пользователя просили представиться. Выведи в ответ только его имя.""")
-                        # self.username = self.processor.llm.process_prompt(user_answer, user_name="user")
-                        # print(f"System concluded: {user_answer}")
-                        self.processor.set_user(self.username)
-                        output_file_name = self.processor.tts.get_audio(f"{self.username}, приятно познакомиться", output_name=target_name)
+            if len(sorted_files):
+                t = time.time_ns()
+                for input_file_name in sorted_files:
+                    # process in place
+                    _name = os.path.split(input_file_name)[-1]
+                    # use same names in output as in input
+                    target_name = os.path.join(self.processor.tts.folder, os.path.splitext(_name)[0] + ".wav")
+                    if _name == "newuser":
+                        # initiate new protocol, ask user his name
+                        self.username = None
+                        output_file_name = self.processor.tts.get_audio("Здравствуйте, меня зовут Сергей Капица! Представьтесь, пожалуйста.", output_name=target_name)
                     else:
-                        async for output_file_name in self.processor.async_process(user_name=self.username, file_to_process=input_file_name, output_name=target_name):
-                            print(output_file_name)
-                os.remove(input_file_name)
-
-            t = time.time_ns() - t
-            if t < self.check_freq:
-                time.sleep(self.check_freq - t)
+                        if self.username is None: # expect user name as an answer
+                            user_answer = self.processor.asr.get_text(input_file_name)
+                            if user_answer is None:
+                                print(f"Couldn't fetch username: {user_answer} from file {input_file_name}. Setting name to Дорогой друг")
+                                self.username = "Дорогой друг"
+                            else:
+                                user_answer = user_answer.strip(".,! ").capitalize()
+                                print(f"User answered: {user_answer}")
+                                self.username = user_answer
+                            # self.processor.llm.set_engine(user_name=None, reset=True, custom_system_prompt="""Ты - система аутентификации для ASR. Пользователя просили представиться. Выведи в ответ только его имя.""")
+                            # self.username = self.processor.llm.process_prompt(user_answer, user_name="user")
+                            # print(f"System concluded: {user_answer}")
+                            self.processor.set_user(self.username)
+                            output_file_name = self.processor.tts.get_audio(f"{self.username}, приятно познакомиться", output_name=target_name)
+                        else:
+                            async for output_file_name in self.processor.async_process(user_name=self.username, file_to_process=input_file_name, output_name=target_name):
+                                print(output_file_name)
+                    os.remove(input_file_name)
+                t = time.time_ns() - t
+                print(f"Processing {input_file_name} took {t/1_000_000_000} s")
+            else:
+                time.sleep(self.check_freq)
 
 if __name__ == "__main__":
     dotenv.load_dotenv()
-    # cleanup .generated
+    
     input_folder = ".received"
     output_folder = ".generated"
     # model_url = "https://huggingface.co/kzipa/kap34_8_8_10/resolve/main/kap34_8_8_10.Q4_K_M.gguf?download=true"
